Refactored audio package
