+++ FredBoat/src/main/java/fredboat/command/music/control/SelectCommand.java
@@ -28,7 +28,7 @@ package fredboat.command.music.control;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import fredboat.audio.player.GuildPlayer;
 import fredboat.audio.player.PlayerRegistry;
-import fredboat.audio.VideoSelection;
+import fredboat.audio.player.VideoSelection;
 import fredboat.audio.queue.AudioTrackContext;
 import fredboat.commandmeta.abs.Command;
 import fredboat.commandmeta.abs.ICommandRestricted;
